All in the main processor:
Removed the use of the log file.  The log file was not being consistently used.  Messages, warnings, and errors were written to the log file, standard output, or both.  And, more and more, they were being output to standard output (sometimes via the common library).  Now everything is written to standard output simplifying the code.  This is consistent with the preprocessor and the shared code in the common library.  Messages, warnings, and errors can be saved by redirecting standard output to a file.

Homogenized the error messages.

Fixed some inconsistencies with how error status was dealt with which lead to fatal errors not being caught and causing a crash later on.