Improved command line and driver file support for the pre-processor using the parsing module in the common library including support for comments and optional arguments/fields and better error handling.  There are a few other details for which I haven't figured the cleanest way to implement so consider this the first of two commits.  This is all backward compatible so nothing should change for anyone's usage.  Backward compatibility was a top priority here.
Also added command line/driver file options to the pre-processor to specify the number of channels and the channel IDs to process.  Again, these are part of the new optional command-line-argument/driver-file-field support so will not affect current usage.  The default channels to process are the legacy Cloud CCI channels as before.