Replacing Ctrl%process_one_phase_only with Ctrl%Types_to_process. If you previously set Ctrl%process_one_phase_only=false please now use,    Ctrl%NTypes_to_process=10    Ctrl%Types_to_process=0,1,2,3,4,5,6,7,8,9
The preprocessor produces the field cldtype, which classifies the type of cloud
in each pixel into one of 10 types. The Types_to_process array lists the flags
in this array that ORAC is allowed to process. If not set, the previous default
behavior is replicated (WAT evalutes 2,3,4,5,9; ICE 5,6,7,8,9; 0-10 otherwise).

The constants defined in constants_cloud_typing_pavolonis can be used for human-
readable input. The intention is that the cldtype field will eventually combined
with aerosol typing in the Ctrl%Type array.