1) Add support for optional driver file arguments to the post processor.
2) Add use_baysian_selection as an optional post processor driver file
argument.  Default is use_baysian_selection=false.  With this option any number
of input primary/secondary file pairs is supported.  To support this lines
after the required lines that do not parse as a label=value are considered a
primary file name followed by a secondary file name on the next line.