Forward model fixes and improvements:
Remove i_equation 1 (previously used for cloud) as an invalid derivation.  Shift equations 2-3 to be 1-2.  Add a new equation 3 to be the same as equation 4 but sun normalized like equation 1 whereas equation 4 is still sun normalized like equation 3.  See detailed comments which also indicate that equations 2 and 4 are incorrectly sun normalized and will be removed in the future.

Fixed the reciprocity obeying equations (now 3 and 4) by using the T_0d (TFBD) LUT in place of the incorrectly produced T_dv LUT (TD).  This is valid as T_dv(theta) = T_0d(theta).

Change the default equation (used for cloud) to 3, the reciprocity obeying equation correctly scaled as equation 1 (old equation 2).

Shift the default for aerosol to equation 1 (old equation 2) resulting in currently no change for the aerosol retrieval.

See new detailed comments in the code.