Add functionality for Ctrl%RS%RsSelm and Ctrl%RS%SRsSelm to the Swansea surface to set the first guess for the S parameters. This mimicks their meaning for the Oxford surface treatment when Ctrl%AP = SelmAux.
For RsSelm,
SelmCtrl) Use the values in Ctrl%RS%b, which can differ over land and sea. The
   uncertainty is a constant fraction of that value, set by Ctrl%RS%Sb.
SelmMeas) If an AOT, REF, P, and gamma are assumed, the Swansea forward model
   can be algebraically inverted for S. This setting does so using the nadir
   radiances and assumptions given by Ctrl%SS_Xb. Their uncertainties are given
   by Ctrl%SS_Sx. They default to Ctrl%Xb and Ctrl%Sx, respectively.
SelmAux) Assume that rho_DD = g * s / (1 - (1-g) * s), so that a first guess for
   S can be drawn from the preprocessed albedo.

For SRsSelm,
SelmCtrl) Use the value in Ctrl%Sx. If Ctrl%RS%add_fractional is true, increase
   this by a fraction of the S value, set by Ctrl%RS%Sb.
SelmMeas) Propagate uncertainties through the equation used.
SelmAux) With RsSelm = SelmAux, draw the uncertainty in rho_DD from the
   preprocessed file.