When extrapolating the nadir azimuth, ignore the sub-satellite tie point.
This was done at the recommendation of SP following on from work he's done with
SLSTR. At the sub-satellite point, the azimuth angle is poorly defined and the
value given within the L1 files isn't overly useful. This can be seen in the
relative azimuth field - there are often sharp discontinuities near the centre
of the swath. If we ignore the central tie point and instead extrapolate the
left and right segments of the orbit separately, the relative azimuth varies
smoothly with a sharp discontinuity at the centre. This is consistent with the
outputs of MODIS and AVHRR.