Update Python scripts to facilitate single-view aerosol retrievals.
If you use the scripts, some settings have been renamed or deleted so please
read the new help documentation. (Or email Adam.) You should also update your
local_defaults file.

The previous version of orac.py required too many assumptions to be able to
run an arbitrary ORAC retrieval. Now, the phases to be processed are specified
by the -s keyword. Each instance provides a string to define the settings
required for that phase (as if called directly). For example,
    orac.py -c 1 2 3 4 5 6 -s "--phase WAT" -s "--phase A70 -u 1 2 3" /some/file
would run a water cloud retrieval using channels 1-6 (as it defaulted to using
all available channels) and an aerosol retrieval (type 70) on channels 1-3
(specified by the -u keyword).

Default retrievals configurations are defined in the local_defaults. These can
be called by name using the -S keyword. local_defaults.py.example defines cloud
and aerosol retrievals for the most common sensors. For example,
    orac.py -S MODIS_C /some/file
would run the default cloud retrieval on a MODIS file. The names are formatted
INSTRUMENT_? where ? is C for cloud, A for aerosol, and J for joint. Default
channels for each sensor are also defined (so -c isn't a required keyword).
Using the regression tests, a given letter can be appended to all tests using
the -T keyword (replacing functionality of the previous -A).

It is also possible to define a file that specifies these settings, using the
--settings_file keyword. Each line of the file defines a phase (equivalent to
a -s statement).