When the retrieval fails, write the status code into variables_retrieved.
Currently, the only way to identify the failure-mode for a pixel is to
compile ORAC with -DDEBUG and manually read the log file. Now, the status
code is written to variables_retrieved, which is output to NCDF.

I am aware that this "breaks" the bit-mask quality of variables_retrieved
but this seemed better than adding another output to already very full files.
I selected variables_retrieved as it's the quality assurance output least
likely to be used by someone that isn't a developer. Failed retrievals are
identified by a fill_value in qcflag and niter.

Status values are defined in ORACConstants.F90.