Add ability to use climatology of Swansea surface parameters.
The preprocessor now accepts the argument Use_Swansea_Climatology. When
true, rather than open the MCD43C1 and C3 files to determine the surface
reflectance, a climatology of the s and p parameters (generated by SP)
is opened and translated into an effective BRDF. The gamma used in that
calculation defaults to 0.3 but can be changed through the Swansea_Gamma
argument in the driver file.

The climatology can be used as prior for the aerosol retrieval by setting
FG and AP to SelmAux in the main processor for ISP and ISS, while
Ctrl%RS%read_full_brdf is true.