Add a small program to determine the version number of RTTOV.
The version of RTTOV used is stored in the header of ORAC files. The Python
scripts previously determined this by looking at the name of the RTTOV library
linked against. That only lists the primary version number.

This update add a small executable that uses the rttov_const module to print
the full version number of the library used. This is called by the Python
scripts and included in the header.

Note that RTTOV 12.2 didn't increment the version number, so this method doesn't
work perfectly there.