Fix use of CAMEL surface emissivity, updating logic of _date_back_search().
_date_back_search() previously would search backwards in time for a valid file,
stepping in days. If other files from that year didn't exist, it would instead
try to find a climatology file with year XXXX.

Now, it can step back in any defined interval, which is useful if you want to
use the previous year to represent this one. Also, the climatology is only tried
once a date search has failed.

The CAMEL emissivity implementation previously failed to find the appropriate
file to open. That logic has been improved.