Read Level 1 land/sea mask in read_slstr().
This is, for some reason, stored in the confidence_VIEW field of the
flags_VIEW.nc file. Unlike other sensors, it is a bitmask so a single
pixel can have multiple classifications. I've decided to classify
coastlines and ocean as "sea" and inland or tidal waters as "land" to
be consistent with the MODIS reader.