start,end values are dealt with in orac_preproc, not read_dimensions()
An exception is made for SEVIRI, where a single image may be tiled
over many files. There, it is useful to maintain a single coordinate
system over those files. To do so, the default start,end values are
set within the read_seviri_dimensions().

Other sensors have x/y coordinates defined for each file, so it is not
sensible to set the start,end variables there.