Fix c5be3ff551766e0ec0695c716685eaea158e2a86
When subsetting an ATSR orbit, there is no need to extrapolate the
y-coordinate for the centre of the final bin as we can read the
oordinate of the next bin along. While the previous fix was correct
to shift when we extrapolate the final pixel, that extrapolation
isn't always appropriate.

This caused the lat/lon of the last row of any subsetted or split
orbit to differ from a full processing by around 1e-3, which was
enough to affect all the other interpolation fields.