Change time interval on backwards searches to years.
_date_back_search() steps backwards in time over a given interval to
find the previous relevant file. In 2b9c98c6e6d01bd6e9581ad9646f4092845a45fa,
I selected an interval of 'months' for surface files. This should be
'years', e.g. if we can't find a file for the month of this observation,
we should use the file from the same month in a previous year, not the
previous month in the same year.

This is more consistent with our approach to climatology files.