Continue preprocessing when there are no surface pixels to calculate.
In a9572a0ea5957c380eeed84acd69e405901b1792, Simon stopped calculation of BRDF
parameters at night. That's fine, but within get_surface_reflectance()
there is a check that >0 pixels need to be processed which prevented
the processing of a night-only scene.

The check now merely returns the function. I've kept the warning in verbose
mode as I see this error a lot in day-to-day work.