Use solar azimuth within the Cox and Munk scheme.
When Greg optimised the Cox-Munk code, he ensured it was consistent
with our previous implimentation. That has left solaz unallocated
ever since the pre-processing was first converted to Fortran. That
does not appear to have been on purpose. As we use to anisotropic
parameterisation of the surface gradient, the solar azimuth will be
relevant.