Change identification of Cloud CCI processing.
Firstly, one no longer uses --phases to control ORAC processing (instead
we use --settings in almost all cases) as this was an ineffective way
of determining if Bayesian type selection should be used.
Secondly, we can no longer assume the LUTs will be called WAT and ICE.
Thirdly, this does require that the processing be in the order water
then ice but I'm hoping the post-processor will be overhauled one day.