Make phase switching default behaviour.
THIS IS RARELY USED as the code only considers this option when Bayesian
type selection is off. That will only be the case, for these scripts,
when processing only water and ice cloud. In that case, we should be
using the phase switching as that's the preferred behaviour for CCI.