Remove the nadir Swansea angular term from the state vector.
The Swansea surface reflectance model has one term for each wavelength
and one for each viewing angle. In the current retrieval, the prior for
the nadir angular term is 0.01 (while the oblique is 0.5) because it
effectively controls the magnitude of the surface reflectance, which can
also be done through the wavelength terms.

In practice, the value was rarely altered by the retrieval as it considered
that term poorly constrained, such that there was a degree of freedom for
noise in every retrieval using this surface model.

This commit removes ISP(1) from the default requested state vector.
This ignores the term entirely. If we want to account for parameter
uncertainty due to it, we'd instead set active=.false. on the appropriate
call to Add_to_State_Vector() in Get_Indexing() but I don't properly
understand uncertainty on this term  and so prefer to simply treat it as
an imposed parameter.