Guess SLSTR start time when processing filenames
SLSTR files are a tad difficult to deal with because the time in
the L1 file can differ from that which ORAC would use. The Python
code fixes this by checking the L1 file for the correct value, but
if that file isn't available it currently returns no time which
complicates handling ORAC part files.

This commit uses whatever timestamp was available. That will always
work when dealing with ORAC outputs and will mostly work with L1
files (and the existing code should work in that case).