Introduce quadratic measurement uncertainty model
Thoroughly characterised sensors can use a more general model of
measurement uncertainty than a single SNR. In the NCDF LUTs, this
is stored in the terms rua, rub, ruc and is currently only available
for the ATSR series of instruments. This commit allows those LUTs
to be opened and used.