Replace calculation of uncertainty in mixed channels
Previously, the uncertainty in BT for a mixed channel was determined
by multiplying the solar constant by NeHomog|Coreg, dividing by dR_dT
from Planck's law, then adding in quadrature to the NeHomog|Coreg terms
from the thermal. This resulted in large uncertainties for very cold
clouds as the observed radiance is O(1%) of the solar constant.

If that approach was correct, it would have been necessary to multiply
the solar constant by cos(solzen).

Instead, I follow the example of the visible channels and treat NeHomog
as a % uncertainty on the observed signal, here represented by Rad
output by the Planck calculation.

The previous behaviour can be generated by setting
USE_OLD_MIXED_UNCERTAINTY to true before compilation.