Use absolute values, rather than log, when setting X in the driver file
In the text LUTs, the optical depth axis is logarithmically spaced and
the values written in the tables are the log of the actual value (e.g.
the first COT value is -20). As such, all references to optical depth
in the state vector must also be in log space (e.g. the prior for
aerosol is -1.5). The effective radius is logarithmically spaced for
aerosol tables but linearly spaced for cloud. This is confusing.

The NCDF LUTs specify if axes are logarithmically or linearly spaced
as an attribute. This commit adds a flag to the LUT Grid structure
indicating if an axis is logarithmic (which is detected by negative
values in the text tables). The set values are then log-ed as required.